% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnBarChartScore.r
\name{fnBarChartScore}
\alias{fnBarChartScore}
\title{Produces a bar chart of score distributions.}
\usage{
fnBarChartScore(data, "scoreVar",  maxScore, "outcomeVar")
}
\arguments{
\item{data}{A data frame containing the scores to be plotted.}

\item{scoreVar}{The name of the specific column containing the scores (numeric or integer). Must be in quotation marks.}

\item{maxScore}{The maximum score possible on the assessment.}

\item{outcomeVar}{The name of the column containing the outcome grades if present. Can be a factor or character string. Must be in quotation marks. Leave this empty if no outcome grouping is required.}
}
\description{
This function plots the distribution of a column of scores, colouring the bars by outcome grade if data is available.
}
\examples{

dataExample<-dataExample
dataExample$outcomePFE[dataExample$Score.Restricted>= 24] <- "Excellent"
dataExample$outcomePFE[dataExample$Score.Restricted>= 13 &  dataExample$Score.Restricted < 24] <- "Pass"  
dataExample$outcomePFE[dataExample$Score.Restricted< 13 ] <- "Fail"
fnBarChartScore(dataExample, "Score.Restricted",  25, "outcomePFE")   


dataExample<-dataExample
dataExample$outcomePF[dataExample$Score.Restricted>= 13] <- "Pass"
dataExample$outcomePF[dataExample$Score.Restricted< 13 ] <- "Fail"
fnBarChartScore(dataExample, "Score.Restricted", 28, "outcomePF")   


dataExample<-dataExample
fnBarChartScore(dataExample, "Score.Restricted", 30) 


}
\author{
Lizzi Gabe-Thomas (elizabeth.gabe-thomas@plymouth.ac.uk)
}
