% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnBoxplot.R
\name{fnBoxplot}
\alias{fnBoxplot}
\title{Produces a score boxplot by grade, stage, assessor, programme (for BDS & DTH only), station/substation
(ISCE), gender, ethnicity or disability. Two variables can be specified to produce split boxplots
within each group.}
\usage{
fnBoxplot(df, xName=NULL, yName=NULL, maxScore=100, forceScheme=NULL, fillBy=NULL)
}
\arguments{
\item{df}{The name of the data frame.}

\item{xName}{The "group" column name (must be in quotation marks) or column number within the data frame.}

\item{yName}{The "score" column name (must be in quotation marks) or column number within the data frame.}

\item{maxScore}{Optional (default value is 100). The maximum possible score available.}

\item{forceScheme}{Optional. Use when the group scheme determined is "USE" but "UBSE" is required,
"PF" but "PFE" is required or "Station" but "Substation" is required (type forceScheme="Y").}

\item{fillBy}{The (optional) "fillBy" column name (must be in quotation marks) or column number within the
data frame when wanting to produce split boxplots (by a second variable).}
}
\description{
This function takes a group column and a score column (and an optional fill 
column) and outputs a ggplot2 object (score boxplot) which can then be customised. The y-axis will default
to a value of 100 but this can be amended (use maxScore=). For Stage and Assessor
groups the data needs to be numeric/integer values; for all other groups the data needs to
be character strings (e.g. "Stn1" or "Sub2A" for ISCE stations/substations). The fillBy parameter does
not operate for station or substation data.
}
\examples{
Basic usage: Examples use dataExample in psychometricsPSMD package. 

fnBoxplot(dataExample,"Grade.UBSE","Score.Perc")

fnBoxplot(dataExample,"Grade.UBSE","Score.Perc",fillBy="Gender")

fnBoxplot(dataExample,"Ethnicity","Score.Restricted",25)

fnBoxplot(dataExample,"Ethnicity","Score.Restricted",maxScore=25,fillBy="Gender")

fnBoxplot(dataExample,5,7,125)

fnBoxplot(dataExample,5,7,10,125)

fnBoxplot(dataExample,"Grade.PF","Score.Perc",forceScheme="Y")

fnBoxplot(dataExample,"Stage","Score.Restricted",25,fillBy="Grade.PF")

}
