% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnCompoundBar.R
\name{fnCompoundBar}
\alias{fnCompoundBar}
\title{Produces a compound bar chart by grade, stage (for PTs and IDS), assessor, programme (for BDS & DTH only), station/substn
(ISCEs), gender, ethnicity or disability.}
\usage{
fnCompoundBar(df, xName=NULL, yName=NULL, percent="Yes", forceSchemeX=NULL, forceSchemeY=NULL)
}
\arguments{
\item{df}{The name of the data frame.}

\item{xName}{The column name to group by on the x-axis. Provide column number or name (in quotation marks if column name).}

\item{yName}{The column name to split the bars by. Provide column number or name (in quotation marks if column name).}

\item{percent}{Defaults to "Yes" for percentage proportion bars; change to percent="No" for frequency bars.}

\item{incl.N}{By default the sample size will show at the base of each bar where not all sizes are equal. These can be removed
by adding incl.N="No".}

\item{forceSchemeX/forceSchemeY}{Optional. For use when the group scheme determined is incorrect. For "USE" but "UBSE" is required,
"PF" but "PFE" is required, "CIN" but "CN", or Station/Stage but Stage/Station is required, add forceSchemeX/Y="Y".
For Assessor/Stage/Station but Domain is required, add forceSchemeX/Y="D".}
}
\description{
This function takes two columns of data and produces a compound bar chart with each
bar (for variable 1) split by percentage proportions (or frequency values) for variable 2.
}
\note{
This function uses \code{\link[ggplot2]{ggplot2}} for the graphics and outputs a ggplot object which can be
customised.
}
\examples{
Basic usage: Examples use dataExample in psychometricsPSMD package. 

fnCompoundBar(dataExample,"Stage","Grade.UBSE")

fnCompoundBar(dataExample,"Grade.PF","Gender",percent="No")

fnCompoundBar(dataExample,"Assessor.ID","Grade.CNINC")

fnCompoundBar(dataExample,5,11,forceSchemeY="UBSE")

fnCompoundBar(dataExample,"Grade.PF","Disability","No",forceSchemeX="PFE")

}
\author{
Jo Cockerill, \email{jo.cockerill@plymouth.ac.uk}
}
