% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnCumFreqFacility.R
\name{fnCumFreqFacility}
\alias{fnCumFreqFacility}
\title{Produces a cumulative frequency plot of facilities by Stage.}
\usage{
fnCumFreqFacility(df,stage=NULL)
}
\arguments{
\item{df}{The name of the data frame containing a facility column for each stage and
a row for each included item). The data frame should NOT include the removed items.}

\item{stage}{This needs to be entered if there is only one stage (default=NULL).}
}
\description{
This function takes a data frame with a column of facilities for each stage group, 
and a row for each included item, and outputs a ggplot2 object which can then be customised. The stage
number only needs to be added where the plot is just for one stage. NEEDS ggplot2, plyr, stats,
stringr and tidyr packages.
}
\examples{
df1<-data.frame(Item=1:100,Y1fac=runif(100),Y2fac=runif(100),Y3fac=runif(100),
Y4fac=runif(100),Y5fac=runif(100)) 
df2<-data.frame(Item=1:100,Y1facility=runif(100))

fnCumFreqFacility(df1)
fnCumFreqFacility(df2)
fnCumFreqFacility(df2,stage=1)

}
\author{
Jo Cockerill, \email{jo.cockerill@plymouth.ac.uk}
}
