% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnDetermineScoreColumns.r
\name{fnDetermineScoreColumns}
\alias{fnDetermineScoreColumns}
\title{A function to determine which column or columns in a data frame contain (numeric) scores}
\usage{
fnDetermineScoreColumns(Dataframe)
}
\arguments{
\item{Dataframe}{Dataframe is a data frame of which you want to determine the possible score column(s).}
}
\description{
fnDetermineScoreColumns is designed to determine which column or columns in a data frame contain (numeric) scores. It operates based on a series of conditionals which evaluate aspects such as whether the variable is numeric, whether the values have a range typical of scoring schemes, or have ranges narrower or broader than scoring schemes used in PSMD.
}
\note{
DZ250417: This is currently an experimental function and will be updated to handle real data more accurately.
DZ250417: Build in a probabilistic element to provide the single best-guess if length(Possible.Score.Columns)>1.
}
\examples{
 
fnDetermineScoreColumns.(dataExample) will return a list of column names for the columns the function thinks could be scores.
In the case of dataExamples, this is multiple possibilities. In reality, there are likely to be fewer. 

}
\author{
Dr Daniel Zahra, \email{daniel.zahra@plymouth.ac.uk}
}
