% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnEmMeans.R
\name{fnEmMeans}
\alias{fnEmMeans}
\title{Produces a table of estimated marginal means covering each level of each factor in a linear model.}
\source{
Written by Martin Roberts (psychometrics@plymouth.ac.uk)
}
\usage{
fnEmMeans<-function(lm,digits=2)
}
\arguments{
\item{lm}{A linear model of class lm, lme, aov, etc.
See \code{\link[emmeans:emmeans]{emmeans help page}} for the range of supported model objects.
The model must contain main effects only: inclusion of interactions terms will result in an error.}

\item{digits}{Number of decimal places required for the estimated marginal means. Defaults to 2.}
}
\value{
The function outputs a data.frame tabulating the sample size and the estimated marginal mean for each factor level..
}
\description{
fnEmMeans identifies the factor levels in a linear model and calculates the corresponding estimated marginal means.
The output is a data.frame tabulating the sample size and the estimated marginal mean for each factor level.
See Examples for more details.
}
\note{
This function uses \code{\link[emmeans]{emmeans}} for the underlying calculations.
}
\examples{
fit1<-aov(breaks ~ wool + tension, warpbreaks)
fnEmMeans(lm=fit1,digits=3)

fnEmMeans(lm=aov(yield ~ block + N + P + K, npk),digits=1)

# Note that factors may need to be defined
fnEmMeans(lm=aov(uptake ~ Type + Treatment + conc , CO2))
   CO2$Concentration<-factor(CO2$conc)
fnEmMeans(lm=aov(uptake ~ Type + Treatment + Concentration , CO2))

}
