% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnGradeRange.R
\name{fnGradeRange}
\alias{fnGradeRange}
\title{Produces a floating barplot of the percentage grade ranges by stage for a MCQ/PT.}
\usage{
fnGradeRange(df1, df2=NULL, grades="UBSE")
}
\arguments{
\item{df1}{The name of the data frame (e.g. the Grade Boundaries table) with the stages as
columns. Column 1 to contain the name of the data within each row. Can include
rows for the min and max scores.}

\item{df2}{The name of the dataframe with the min and max scores (e.g. the Descriptive Stats
table) if not present in df1. Stages must be as columns and df2 must have the same
number of columns as df1. Column 1 to contain the name of the data within each row.}

\item{grades}{Specify the grade group, in inverted commas ("UBSE" (dafault) or "US").}
}
\description{
This function takes one or two dfs of grade boundaries and min/max scores
and produces a grade range barplot as a ggplot2 object.
}
\examples{
testData1<-data.frame(Value=c("Minimum","Maximum","Borderline boundary",
"Satisfactory boundary","Excellent boundary"),Y1=c(0.0,18.2,1.4,2.6,15),
Y2=c(9.2,59.6,12.3,15,35.4),Y3=c(18.2,55.8,20.9,29.1,51.3),
Y4=c(29.4,72,30.1,34.6,58),Y5=c(38,72,NA,45.09,NA))

testData2<-data.frame(Value=c("Minimum","Maximum","Borderline boundary",
"Satisfactory boundary","Excellent boundary"), Y2=c(1.25,54,-4.2,7.12,38.23),
Y3=c(36,72,22.6,34.14,62.84),Y4=c(44.75,78.25,34.26,45.45,69.06),
Y5=c(45.25,85.2,39.9,52.2,75.3))

testData3<-data.frame(Value=c("Minimum","Maximum","Satisfactory boundary"),
Y1=c(20.93,49.80,25.1),Y2=c(41.8,68.8,49.2))

fnGradeRange(testData1)

fnGradeRange(testData2)

fnGradeRange(testData3,grades="US")

}
