% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnHeatmap.R
\name{fnHeatmap}
\alias{fnHeatmap}
\title{Produces a heatmap of the skills and subjects for the BMBS Progress Tests.}
\usage{
fnHeatmap(df, colSkill="Skill", colSubject="Subject", itemsRemoved=NULL, zerosBlank="Yes")
}
\arguments{
\item{df}{The name of the answer key data frame in item number order.}

\item{colSkill}{The Skill column name (must be in quotation marks) or column 
number within the data frame (default colname is "Skill").This should contain
the Skills as uppercase values.}

\item{colSubject}{The Subject column name (must be in quotation marks) or 
column number within the data frame (default colname is "Subject"). This should
contain the Subjects as numeric values.}

\item{itemsRemoved}{Use if any items have been removed (optional). A vector
of the item numbers which have been removed from the test.}

\item{zerosBlank}{Defaults to "Yes" so zero values do not show in the heatmap tiles.
Add parameter zerosBlank="No" for zero values to show.}
}
\description{
This function takes the Answer Key dataframe, the skill and
subject columns along with any removed items and produces a heatmap in ggplot2.
This function requires ggplot2, plyr and reshape packages.
}
\examples{
Item<-seq(1:20)
Skill<-c("G","G","H","I","F","S","T","G","G","M","G","G","H","I","F","S","F","G","G","M")
Subject<-c(40,2030,1080,71,20,330,1110,2300,1500,1275,970,370,1110,570,1510,340,460,450,50,460)
answerKey<-data.frame(Item,Skill,Subject) 
fnHeatmap(answerKey)
fnHeatmap(answerKey, itemsRemoved=c(2,12))
fnHeatmap(answerKey, zerosBlank="No")

}
