% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnHofstee.R
\name{fnHofstee}
\alias{fnHofstee}
\title{Carries out the Hofstee standard setting method and displays the method and the result on a graph.}
\source{
Written by Martin Roberts (psychometrics@plymouth.ac.uk)
}
\usage{
fnHofstee(x, pars=c(62.5,75,0,20), cutDigits=2)
}
\arguments{
\item{x}{A vector of percentage scores.}

\item{pars}{A vector of length 4 containing the Hofstee parameters in this order: 
"Minimum cut score", "Maximum cut score", "Minimum failure rate", "Maximum failure rate".}

\item{cutDigits}{Number of decimal places to which the Hofstee cut score is displayed on the plot. 
Defaults to 2 but in situations where the cut score is to be used in further calculations (e.g. AD(T)K tests) change this to 4.}
}
\value{
The function exports either 3 or 4 objects into the Global Environment.
 \enumerate{
  \item HofsteePlot: a ggplot object displaying the method and the solution point.
  \item HofsteePoint: a length 2 vector containing the x,y coordinates of the Hofstee solution point.
  \item HofsteeFail: a logical TRUE/FALSE indicating whether the method failed with the supplied parameters.
  \item HofsteeFailText: A character string describing how the parameters have been changed to effect a solution (only created when HofsteeFail=TRUE).
}
}
\description{
fnHofstee plots a cumulative relative frequency curve (ogive) of assessment scores and adds a graphical display of the Hofstee standard setting method.
The output is a ggplot object, which can be saved or customised as needed, together with the coordinates of the Hofstee point.
See Examples for more details.
}
\note{
This function uses \code{\link[ggplot2]{ggplot2}} for the underlying graphics and \code{\link[spatstat]{spatstat}} to calculate the Hofstee intersection point.
}
\examples{
fnHofstee(x=rnorm(90,75,7))
fnHofstee(x=rnorm(90,60,10),pars=c(45,65,0,20),cutDigits=4)
fnHofstee(rnorm(90,50,10),c(45,60,0,10))
fnHofstee(rnorm(90,80,5),c(50,60,5,20))

}
