% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnNumToWord.R
\name{fnNumToWord}
\alias{fnNumToWord}
\title{Changes a whole number or integer value (up to 999) to a lower case character string,
with the option to convert to title case.}
\usage{
fnNumToWord(x, Capital=FALSE)
}
\arguments{
\item{x}{The value or object to convert.}

\item{Capital}{Default is FALSE. Option to create a title case character string with Capital=TRUE.}
}
\description{
This function takes an object, integer or whole number (up to a value of 999/"999") and will convert
it to a character string. Useful when including objects (e.g number of passing students) within report text. 
Where the object or value is greater than 999/"999" and/or includes a decimal point the original string/
numeric value will be returned.
}
\examples{
x<-3
fnNumToWord(x)
 
fnNumToWord(12,TRUE)

fnNumToWord(999)

fnNumToWord(10.1)
 
studentsSick<-5
paste0("There were ",fnNumToWord(studentsSick)," students absent due to sickness.")

studentsECs<-8
paste0(fnNumToWord(studentsECs,TRUE)," students were absent with ECs.")

studentsECs<-0
paste0("There were ",fnNumToWord(studentsECs)," students with ECs.")

fnNumToWord(999,TRUE)

}
