% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnPlotBReg.R
\name{fnPlotBReg}
\alias{fnPlotBReg}
\title{Plot of the Borderline Regression method.}
\source{
Written by Martin Roberts (psychometrics@plymouth.ac.uk)
}
\usage{
fnPlotBReg(x,gradeCol="Grade",scoreCol="Score")
}
\arguments{
\item{x}{A data frame containing the global grade and score data.}

\item{gradeCol}{Name of the column containing the global grades.}

\item{scoreCol}{Name of the column containing the assessment scores.}
}
\value{
A ggplot object which can be saved or customised as needed. 
Additionally the linear regression model (bRegLine) and the pass mark (passMark) are passed to the global environment.
}
\description{
fnPlotBReg creates a scatterplot assessment scores against global grades.
A regression line is added to the plot and the intersection of this line with the 'borderline group' is used to calculate a pass mark.
The main output from fnPlotBReg is a ggplot object which can then be saved or customised as needed.
Additionally the function passes the linear regression model and the pass mark into the global environment (see Examples).
}
\note{
This function uses \code{\link[ggplot2]{ggplot2}} for the underlying graphics,
and uses the str_wrap function from the \code{\link[stringr]{stringr}} package to wrap the x axis labels.
}
\examples{
library("ggplot2")
Data<-data.frame(Grade = factor(c(rep("Unsatisfactory",5),rep("Borderline",10),rep("Low Satisfactory",20),
       rep("High Satisfactory",25),rep("Excellent",10)),ordered=TRUE,
       levels = c("Unsatisfactory","Borderline","Low Satisfactory","High Satisfactory","Excellent")),
       Score = c(5,8:10,10, 10:14,11:15, 12:16,13:17,13:17,14:18, 13:17,14:18,15:19,16:20,17:21, 18:22,20:23,25),
       Grade4 = factor(c(rep("Unsatisfactory",5),rep("Borderline",20),rep("Satisfactory",32),rep("Excellent",13)),
       ordered = TRUE, levels = c("Unsatisfactory","Borderline","Satisfactory","Excellent")),
       Score4 = c(5,6,8:10, 9:13,10:14, 10:14,11:15,11:15,12:16, 12:16,13:17,13:17,14:18,14:18,15:20, 18:20,20))

fnPlotBReg(Data)
bRegLine  # Details of the regression model
summary(bRegLine)   # Summary of the regression model
car::outlierTest(bRegLine)   # Test for significant outliers in the model
passMark  # The pass mark obtained from the model

fnPlotBReg(Data,"Grade4","Score4")
	

}
