% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnPlotHistoricStats.r
\name{fnPlotHistoricStats}
\alias{fnPlotHistoricStats}
\title{Time series plot of progress test statistics against test number, grouped by cohort}
\source{
Written by Martin Roberts (psychometrics@plymouth.ac.uk)
}
\usage{
fnPlotHistoricStats<-function(x,statistic,lo=NULL,hi=NULL,Ntests=24)
}
\arguments{
\item{x}{A data frame containing the historic progress test statistics.}

\item{statistic}{Name of the statistic to be plotted. Must be one of the following: 
"Correct","Incorrect","DontKnow","Mean","Min","Max","SD","Alpha","Phi","RelativeSEM","AbsoluteSEM","Borderline","Satisfactory","Excellent".}

\item{lo}{Minimum value for plotting the y axis (optional). If specified will overide the automatic calculation of the lower axis limit.}

\item{hi}{Maximum value for plotting the y axis (optional). If specified will overide the automatic calculation of the upper axis limit.}

\item{Ntests}{Number of tests (counting from the most recent) that are included in the plot (default = 24). Avoids creating confused plots when the number of tests is large.}
}
\value{
A ggplot object which can be saved or customised as needed.
}
\description{
fnPlotHistoricStats creates a time series plot of progress test statistics against test number grouped by cohort.
The output from fnHistogramScore is a ggplot object which can then be saved or customised as needed.
See Examples for more details.
}
\note{
This function uses \code{\link[ggplot2]{ggplot2}} for the underlying graphics.
}
\examples{
library("xlsx")
filename<-c("X:\\\\1718\\\\BMBS\\\\AMK\\\\PT17\\\\Reports\\\\M_1718_PT17_V2 Historic Stats.xlsx",
"X:\\\\1718\\\\BDS\\\\ADK\\\\ADK23\\\\Reports\\\\D_1718_ADK23_V2 Historic Stats.xlsx",
"X:\\\\1718\\\\BScDTH\\\\ADTK\\\\ADTK07\\\\Reports\\\\TH_1718_ADTK07_V2 Historic Stats.xlsx",
"X:\\\\Plymouth\\\\PgDipPAS\\\\AMK\\\\1617\\\\PAPT12\\\\Reports\\\\P_1617_PAPT12_v1.1 Historic Stats.xlsx")
historicStats<-read.xlsx(filename[1],sheetIndex=1,stringsAsFactors=FALSE)

fnPlotHistoricStats(historicStats,"Mean")
fnPlotHistoricStats(historicStats,"Incorrect",lo=0,hi=100)
	
Outputs a ggplot object, which can them be edited as necessary.
	plotExample<-fnPlotHistoricStats(historicStats,"Alpha")
	plotExample  # plot output
	plotExample+ggtitle("Time Series Plot of Cronbach's Alpha") # plot output plus title

}
