% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRound.r
\name{fnRound}
\alias{fnRound}
\title{A function to round values in accordance with PSMD policy}
\usage{
fnRound(x, n)
}
\arguments{
\item{x}{numeric value or vector of values to be rounded. NAs are preserved but non-numeric values will cause an error.}

\item{n}{number of decimal places required (default = 0). Trailing zeroes will be excluded.}
}
\description{
fnRound will round numeric values in accordance with PSMD policy. The policy (as at August 2016) proposes the use of the asymmetric round-half-up method (arithmetic rounding).
This method rounds values at or above the midpoint away from zero and values below the midpoint towards zero to the closest required decimal place. For example, when rounded to two decimal places values of 1.000 to 1.004 become 1.00 and values of 1.005 to 1.009 become 1.01. This applies irrespective of a number's sign so that values of -1.000 to -1.004 become -1.00 and values of -1.005 to -1.009 become -1.01.

Adapted from an anonymous post at Statistically Significant; http://alandgraf.blogspot.co.uk/2012/06/rounding-in-r.html
}
\examples{
 fnRound(x = 13.4345, n = 2)

fnRound(13.4345, 3)

fnRound(c(2.245, -2.245, 0.005, 1.133, NA, -0.515), 2)

fnRound(seq(0.05,1.15,0.1), 1)

}
