% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnScatterplot.R
\name{fnScatterplot}
\alias{fnScatterplot}
\title{Scatterplots for Test-Retest and Facility-by-ResponseRate.}
\usage{
fnScatterplot(Data=NULL, x=NULL, y=NULL, Grouping=NULL, Type=NULL, Items=NULL)
}
\arguments{
\item{Data}{Data should be a dataframe including variables for at least x and y. It should also include Grouping and Items if used as fnScatterplot checks Data for the specified variables. Data should be entered as a named object, i.e. Data.Scores rather than as a string, "Data.Scores".}

\item{x}{Ideally a numeric variable of values to be plotted on the x-axis. Should be entered as a string, i.e. x="variable.name" rather than x=variable.name. fnScatterplot will attempt to coerce non-numeric and factorised variables into suitable formats. Proportions will be converted to percentage.}

\item{y}{Ideally a numeric variable of values to be plotted on the x-axis. Should be entered as a string, i.e. "y=[variable.name]" rather than y=variable.name. fnScatterplot will attempt to coerce non-numeric and factorised variables into suitable formats. Proportions will be converted to percentage.}

\item{Grouping}{The variable by which TRT points should be groups, usully likely to be Stage. Should be entered as a string, i.e. Grouping="Stage" rather than Grouping=variable.name. fnScatterplot will attempt to coerce factorised variables into suitable formats. Can be NULL.}

\item{Type}{The type of plot required. Specify "TRT" for test-retest (with or without Grouping), or "IA" for facility by response rate. If using IA, the Items argument must be specified.}

\item{Items}{The variable which specifies item numbers to be used to label points in a Facility-by-ResponseRate plot.

 @note 
 DZ180118: Error messages are sequential; i.e. if Data is missing, no other argumenrts will be checked. Also, all variables should be in Data; possibly need to add some flexibility here to pull arguments from other dataframes.}
}
\description{
fnScatterplot creates scatterplots for test-retest performance (with or without grouping by Stage), and scatterplots of Facility by Response Rate.

The output from fnScatterplot is a ggplot object which can then be saved or customised as needed.
}
\examples{

Examples of warning combinations and working plots (using Data.1 below; need to update to use dataExample):

Data.1<-data.frame(Item=c(1:20),Stage=factor(c(1,1,1,1,1,2,2,2,2,2,3,3,4,4,4,3,2,4,5,5)), Test1=c(25,10,12,13,25,54,54,44,36,29,65,55,59,57,58,70,75,82,95,81), Test2=c(27,17,20,18,32,60,50,39,30,35,68,62,50,70,65,79,78,90,91,90),ResponseRate=c(25,32,36,84,85,74,65,55,41,52,24,26,96,84,75,77,65,68,63,69),Facility=c(0.10,0.25,0.40,0.90,0.82,0.65,0.84,0.24,0.85,0.74,0.07,0.09,0.98,0.89,0.76,0.74,0.68,0.71,0.80,0.73))

fnScatterplot(Data=NULL, x=NULL, y=NULL, Grouping=NULL, Type=NULL, Items=NULL)
fnScatterplot(Data=Data.1, x=NULL, y=NULL, Grouping=NULL, Type=NULL, Items=NULL)
fnScatterplot(Data=Data.1, x="Test1", y=NULL, Grouping=NULL, Type=NULL, Items=NULL)
fnScatterplot(Data=Data.1, x="Test1", y="Test2", Grouping=NULL, Type="TRT", Items=NULL)
fnScatterplot(Data=Data.1, x="Test1", y="Test2", Grouping=NULL, Type="IA", Items=NULL)
fnScatterplot(Data=Data.1, x="Test1", y="Test2", Grouping=NULL, Type="Expelliarmus", Items=NULL)
fnScatterplot(Data=Data.1, x="Test1", y="Mysterious", Grouping=NULL, Type="IA", Items=NULL)
fnScatterplot(Data=Data.1, x="Test1", y="Mysterious", Grouping=NULL, Type="IA", Items="Item")
fnScatterplot(Data=Data.1, x="Test1", y="Mysterious", Grouping=NULL, Type="TRT", Items=NULL)
fnScatterplot(Data=Data.1, x="Mysterious", y="Test2", Grouping=NULL, Type="TRT", Items=NULL)
fnScatterplot(Data=Data.1, x="Test1", y="Test2", Grouping="Stage", Type="TRT", Items=NULL)
fnScatterplot(Data=Data.1, x="Test1", y="Test2", Grouping=NULL, Type="IA", Items="Mysterious")
fnScatterplot(Data=Data.1, x="Test1", y="Test2", Grouping=NULL, Type="IA", Items="Item")

}
\author{
Dr Daniel Zahra, \email{daniel.zahra@plymouth.ac.uk}
}
