% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnUpdate.r
\name{fnUpdate}
\alias{fnUpdate}
\title{A function to automate updating the psychometricsPSMD package}
\usage{
fnUpdate(PackageName, Reinstall)
}
\arguments{
\item{PackageName}{Can be omitted if you want to update psychometricsPSMD as it defaults to this. Specifying any other name should update the named package, but ssee 'Note' section.}

\item{Reinstall}{Reinstall determines whether the package is reinstalled and reattached regardless of its current status (e.g. When Reinstall==TRUE, the package named in PackageName will be reinstalled and reattached even if it is already installed and/or attached). Set Reinstall=TRUE for updating psychometricsPSMD.}
}
\description{
A function to automate updating the psychometricsPSMD package. .

Adapted from an anonymous post at Statistically Significant; http://alandgraf.blogspot.co.uk/2012/06/rounding-in-r.html
}
\note{
DZ260417: The argument character.only = TRUE works for library(), but possibly not for install.packages. So although PackageName in fnUpdate() could be changed to install and attach any package, it might hit errors at the install.package() phase. PackageName currently defaults to psychometricsPSMD though, so unless specified otherwise, fnUpdate handles psychometricsPSMD.
}
\examples{
 fnUpdate(Reinstall = TRUE) will force a clean install and reattachment of psychometricsPSMD.

}
\author{
Dr Daniel Zahra, \email{daniel.zahra@plymouth.ac.uk}
}
